refactor: #2 랜딩페이지 리팩토링 - 못생김 방지: 최소 너비 1880px 지정 - scrollto 일부 수정 - 파일 구조 일부 수정
