+++ src/app/page.tsx
@@ -1,9 +1,9 @@
 import CardArticle from "@/components/landing/CardArticle";
-import HeadArticle from "@/components/landing/HeadArticle";
+import HeadArticle from "@/components/landing/head-article/HeadArticle";
 import SecondArticle from "@/components/landing/SecondArticle";
 import ServiceArticle from "@/components/landing/ServiceArticle";
 import ThirdArticle from "@/components/landing/ThirdArticle";
-import TopButton from "@/components/landing/TopButton";
+import TopButton from "@/components/landing/head-article/TopButton";
 
 export default function Home() {
   return (
