+++ src/components/landing/ServiceArticle.tsx
@@ -2,7 +2,7 @@ import Support from "../common/Support";
 
 function ServiceArticle() {
   return (
-    <article className="flex h-[1280px] items-center justify-center">
+    <article className="flex h-[1280px] min-w-[1880px] items-center justify-center">
       <div className="flex h-[817px] items-center justify-start gap-[124px]">
         <div className="flex h-[808px] flex-col items-start justify-between">
           <div className="flex flex-col gap-[30px] text-left">
